FInished Styling. Started developing functionality.
The inital styles for the calculator have been completed. There was an issue where the HTML body was the calculator itself - this has been resolved. However, the plan to center the calculator onto the viewport hasn't been fixed - will fix later once functionality has been completed. The app has the ability to pass an arrow function to the Button component and able to console log which button has been clicked.